Allocate always from hugepages.
